Guest reroute

+++ app/controllers/application_controller.rb
@@ -20,19 +20,21 @@ class ApplicationController < ActionController::Base
 
  def after_sign_in_path_for(resource_or_scope)
     case resource_or_scope
-    when :guest, Guest
-      store_location = session[:forwarding_url]
-      session.delete(:forwarding_url)
+    when :guest, Guest 
+      store_location = session[:forwarding_url] || session[:host_page]
+      session.delete(:forwarding_url) if store_location = session[:forwarding_url]
+      session.delete(:host_page) if store_location = session[:host_page]
       (store_location.nil?) ? "/" : store_location.to_s
 
     when :host, Host
-      store_location = session[:forwarding_url]
+      store_location = session[:forwarding_url] 
       session.delete(:forwarding_url)
       (store_location.nil?) ? "/" : store_location.to_s
     else
       super
     end
   end
+
 
 
  	# alias_method :after_sign_in_path_for, :after_sign_up_path_for
